+++ src/superface-test.interfaces.ts
@@ -12,7 +12,6 @@ import {
 } from '@superfaceai/one-sdk/dist/internal/interpreter/variables';
 import { Definition } from 'nock/types';
 
-import { AnalysisResult } from './nock/analyzer';
 import { ErrorCollection } from './nock/matcher.errors';
 
 export interface SuperfaceTestConfigPayload {
@@ -30,7 +29,7 @@ export interface HashOptions {
   testName?: string;
 }
 
-export type AlertFunction = (analysis: AnalysisResult) => void | Promise<void>;
+export type AlertFunction = (report: TestReport) => unknown | Promise<unknown>;
 
 export type SuperfaceTestRun = Omit<
   SuperfaceTestConfigPayload,
@@ -70,7 +69,6 @@ export interface RecordingProcessOptions {
   beforeRecordingLoad?: ProcessingFunction;
   hideInput?: string[];
   recordingVersion?: string;
-  alert?: AlertFunction;
   fullError?: boolean;
 }
 
@@ -95,4 +93,4 @@ export interface TestCoverageBase {
 //       error: PerformError | string;
 //     };
 
-export type TestAnalysis = TestCoverageBase[];
+export type TestReport = TestCoverageBase[];
